refactor : 훅 간편화
